Rename existing BftProtocolSchedule classes to append Builder for clarity Signed-off-by: Simon Dudley <simon.dudley@consensys.net>
